remove old pipeline pathway and files, add API for deprecated pathway
