+++ src/deepsparse/transformers/__init__.py
@@ -120,4 +120,3 @@ _check_transformers_install()
 from .helpers import *
 from .loaders import *
 from .pipelines import *
-from .server import *
